+ merge: Scintilla's current GitHub mirror (404: 0430) development + fix: Deactivate "Focused View": set FoldingLevel to default before applying Lexer's Style (-> speed-up)

+++ scintilla/src/UniConversion.h
@@ -47,7 +47,7 @@ inline bool UTF8IsAscii(int ch) noexcept {
 }
 
 enum { UTF8MaskWidth=0x7, UTF8MaskInvalid=0x8 };
-int UTF8Classify(const unsigned char *us, int len) noexcept;
+int UTF8Classify(const unsigned char *us, size_t len) noexcept;
 
 // Similar to UTF8Classify but returns a length of 1 for invalid bytes
 // instead of setting the invalid flag
